requirements.txt
# Core FastAPI + ASGI server
fastapi==0.115.5
uvicorn[standard]==0.32.0
watchfiles==0.23.0  # Soporte para recarga automática del servidor

# ORM y base de datos
sqlalchemy==2.0.36
# psycopg2-binary==2.9.10  # PostgreSQL
# pyodbc==5.2.0            # SQL Server (descomenta si lo usas)

# Alembic para migraciones
alembic==1.14.0

# Validaciones y esquemas (Pydantic v2, compatible con Python 3.14)
pydantic==2.9.2
pydantic-core==2.23.4

# Seguridad y autenticación JWT
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Autenticación con Google
httpx==0.27.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1

# Envío de correos
sendgrid==6.11.0

# Manejo de formularios y archivos
python-multipart==0.0.9

# Validación de correos y manejo de .env
email-validator==2.2.0
python-dotenv==1.0.1

# Utilidades varias
loguru==0.7.2
python-slugify==8.0.4
arrow==1.3.0
phonenumbers==8.13.40

# Zona horaria (recomendado en entornos Linux/Alpine)
tzdata==2024.2
